+++ src/plugins/intel_npu/src/al/CMakeLists.txt
@@ -26,6 +26,12 @@ target_link_libraries(${TARGET_NAME}
         openvino::runtime::dev
 )
 
+target_include_directories(${TARGET_NAME}
+    PUBLIC
+        $<TARGET_PROPERTY:LevelZero::LevelZero,INTERFACE_INCLUDE_DIRECTORIES>
+        $<TARGET_PROPERTY:LevelZero::NPUExt,INTERFACE_INCLUDE_DIRECTORIES>
+)
+
 set_target_properties(${TARGET_NAME} PROPERTIES INTERPROCEDURAL_OPTIMIZATION_RELEASE ${ENABLE_LTO})
 ov_add_clang_format_target(${TARGET_NAME}_clang FOR_TARGETS ${TARGET_NAME})
