--- src/plugins/intel_npu/src/al/include/icompiler_adapter.hpp
@@ -1,26 +1,0 @@
-// Copyright (C) 2018-2024 Intel Corporation
-// SPDX-License-Identifier: Apache-2.0
-//
-
-// Compiler Interface
-
-#pragma once
-
-#include <memory>
-#include <vector>
-
-#include "igraph.hpp"
-#include "intel_npu/al/config/common.hpp"
-#include "intel_npu/al/icompiler.hpp"
-
-namespace intel_npu {
-
-class ICompilerAdapter {
-public:
-    virtual std::shared_ptr<IGraph> compile(const std::shared_ptr<const ov::Model>& model,
-                                            const Config& config) const = 0;
-    virtual std::shared_ptr<IGraph> parse(const std::vector<uint8_t>& network, const Config& config) const = 0;
-    virtual ov::SupportedOpsMap query(const std::shared_ptr<const ov::Model>& model, const Config& config) const = 0;
-};
-
-}  // namespace intel_npu
