+++ src/plugins/intel_npu/src/backend/CMakeLists.txt
@@ -35,25 +35,21 @@ target_link_libraries(${TARGET_NAME}
 ov_install_static_lib(${TARGET_NAME} ${NPU_INTERNAL_COMPONENT})
 
 if(TARGET ze_loader)
-    if(NOT BUILD_SHARED_LIBS)
-        # Support link of static runtime in case system does not have ze_loader
-        install(TARGETS ze_loader EXPORT OpenVINOTargets
-            RUNTIME DESTINATION ${OV_CPACK_RUNTIMEDIR} COMPONENT ${NPU_PLUGIN_COMPONENT}
-            ARCHIVE DESTINATION ${OV_CPACK_ARCHIVEDIR} COMPONENT ${NPU_PLUGIN_COMPONENT}
-            LIBRARY DESTINATION ${OV_CPACK_LIBRARYDIR} COMPONENT ${NPU_PLUGIN_COMPONENT})
+    install(TARGETS ze_loader EXPORT OpenVINOTargets
+        RUNTIME DESTINATION ${OV_CPACK_RUNTIMEDIR} COMPONENT ${NPU_PLUGIN_COMPONENT}
+        ARCHIVE DESTINATION ${OV_CPACK_ARCHIVEDIR} COMPONENT ${NPU_PLUGIN_COMPONENT}
+        LIBRARY DESTINATION ${OV_CPACK_LIBRARYDIR} COMPONENT ${NPU_PLUGIN_COMPONENT})
 
-        install(TARGETS utils EXPORT OpenVINOTargets
-            RUNTIME DESTINATION ${OV_CPACK_RUNTIMEDIR} COMPONENT ${NPU_PLUGIN_COMPONENT}
-            ARCHIVE DESTINATION ${OV_CPACK_ARCHIVEDIR} COMPONENT ${NPU_PLUGIN_COMPONENT}
-            LIBRARY DESTINATION ${OV_CPACK_LIBRARYDIR} COMPONENT ${NPU_PLUGIN_COMPONENT})
+    install(TARGETS utils EXPORT OpenVINOTargets
+        RUNTIME DESTINATION ${OV_CPACK_RUNTIMEDIR} COMPONENT ${NPU_PLUGIN_COMPONENT}
+        ARCHIVE DESTINATION ${OV_CPACK_ARCHIVEDIR} COMPONENT ${NPU_PLUGIN_COMPONENT}
+        LIBRARY DESTINATION ${OV_CPACK_LIBRARYDIR} COMPONENT ${NPU_PLUGIN_COMPONENT})
 
-        # export to local tree to build against static build tree
-        export(TARGETS ze_loader NAMESPACE openvino::
-            APPEND FILE "${CMAKE_BINARY_DIR}/OpenVINOTargets.cmake")
+    export(TARGETS ze_loader NAMESPACE openvino::
+        APPEND FILE "${CMAKE_BINARY_DIR}/OpenVINOTargets.cmake")
 
-        export(TARGETS utils NAMESPACE openvino::
-            APPEND FILE "${CMAKE_BINARY_DIR}/OpenVINOTargets.cmake")
-    endif()
+    export(TARGETS utils NAMESPACE openvino::
+        APPEND FILE "${CMAKE_BINARY_DIR}/OpenVINOTargets.cmake")
 
     # Support tests to run with ze_loader
     install(TARGETS ze_loader
