+++ src/plugins/intel_npu/src/backend/include/zero_device.hpp
@@ -20,8 +20,6 @@ class ZeroDevice : public IDevice {
 public:
     ZeroDevice(const std::shared_ptr<ZeroInitStructsHolder>& initStructs);
 
-    std::shared_ptr<IExecutor> createExecutor(const std::shared_ptr<IGraph>& graph, const Config& config) override;
-
     std::string getName() const override;
     std::string getFullDeviceName() const override;
     Uuid getUuid() const override;
@@ -37,7 +35,6 @@ public:
 
     std::shared_ptr<SyncInferRequest> createInferRequest(const std::shared_ptr<const ICompiledModel>& compiledModel,
                                                          const std::shared_ptr<IGraph>& graph,
-                                                         const std::shared_ptr<IExecutor>& executor,
                                                          const Config& config) override;
     void updateInfo(const Config& config) override {
         log.setLevel(config.get<LOG_LEVEL>());
