+++ src/plugins/intel_npu/src/backend/include/zero_executor.hpp
@@ -9,7 +9,6 @@
 
 #include <mutex>
 
-#include "igraph.hpp"
 #include "intel_npu/utils/logger/logger.hpp"
 #include "openvino/runtime/properties.hpp"
 #include "zero_init.hpp"
@@ -19,20 +18,18 @@ namespace intel_npu {
 
 class ZeroExecutor final : public IExecutor {
 public:
-    ZeroExecutor(const std::shared_ptr<const ZeroInitStructsHolder>& initStructs,
-                 const std::shared_ptr<IGraph>& graph,
+    ZeroExecutor(ze_graph_handle_t graphHandle,
+                 ze_device_handle_t deviceHandle,
+                 ze_context_handle_t contextHandle,
+                 ze_graph_dditable_ext_curr_t& graphDdiTableExt,
+                 ze_command_queue_npu_dditable_ext_curr_t& commandQueueDdiTable,
                  const Config& config,
-                 uint32_t group_ordinal);
+                 uint32_t groupOrdinal);
 
     ZeroExecutor(const ZeroExecutor&) = delete;
     ZeroExecutor& operator=(const ZeroExecutor&) = delete;
 
-    ~ZeroExecutor() override;
-
-    struct ArgumentDescriptor {
-        ze_graph_argument_properties_3_t info;
-        uint32_t idx;
-    };
+    ~ZeroExecutor() override{};
 
     void setWorkloadType(const ov::WorkloadType workloadType) const override;
     void mutexLock() const;
@@ -41,22 +38,10 @@ public:
     inline const std::shared_ptr<CommandQueue>& getCommandQueue() const {
         return _command_queue;
     }
-    inline const std::vector<ArgumentDescriptor>& get_input_descriptors() const {
-        return _input_descriptors;
-    }
-    inline const std::vector<ArgumentDescriptor>& get_output_descriptors() const {
-        return _output_descriptors;
-    }
 
 private:
     const Config _config;
     Logger _logger;
-
-    const std::shared_ptr<const ZeroInitStructsHolder> _initStructs;
-    std::shared_ptr<IGraph> _graph;
-
-    std::vector<ArgumentDescriptor> _input_descriptors;
-    std::vector<ArgumentDescriptor> _output_descriptors;
 
     std::shared_ptr<CommandQueue> _command_queue;
