+++ src/plugins/intel_npu/src/backend/include/zero_infer_request.hpp
@@ -24,7 +24,6 @@ public:
     explicit ZeroInferRequest(const std::shared_ptr<ZeroInitStructsHolder>& initStructs,
                               const std::shared_ptr<const ICompiledModel>& compiledModel,
                               const std::shared_ptr<IGraph>& graph,
-                              const std::shared_ptr<const IExecutor>& executor,
                               const Config& config,
                               uint32_t group_ordinal);
 
@@ -88,7 +87,6 @@ private:
 
     const std::shared_ptr<ZeroInitStructsHolder> _initStructs;
     const std::shared_ptr<IGraph> _graph;
-    const std::shared_ptr<const IExecutor> _executorPtr;
     const ZeroExecutor* _executor;
     const uint32_t _group_ordinal;
     const Config _config;
