+++ src/plugins/intel_npu/src/backend/src/zero_device.cpp
@@ -91,11 +91,6 @@ ZeroDevice::ZeroDevice(const std::shared_ptr<ZeroInitStructsHolder>& initStructs
     log.debug("ZeroDevice::ZeroDevice - init completed");
 }
 
-std::shared_ptr<IExecutor> ZeroDevice::createExecutor(const std::shared_ptr<IGraph>& graph, const Config& config) {
-    OV_ITT_SCOPED_TASK(itt::domains::LevelZeroBackend, "Device::createExecutor");
-    return std::make_shared<ZeroExecutor>(_initStructs, graph, config, _group_ordinal);
-}
-
 std::string ZeroDevice::getName() const {
 //    KMD is setting usDeviceID from VpuFamilyID.h
 #define NPU_3720_P_DEVICE_ID 0x7D1D
@@ -208,9 +203,8 @@ uint32_t ZeroDevice::getGroupOrdinal() const {
 std::shared_ptr<SyncInferRequest> ZeroDevice::createInferRequest(
     const std::shared_ptr<const ICompiledModel>& compiledModel,
     const std::shared_ptr<IGraph>& graph,
-    const std::shared_ptr<IExecutor>& executor,
     const Config& config) {
-    return std::make_shared<ZeroInferRequest>(_initStructs, compiledModel, graph, executor, config, _group_ordinal);
+    return std::make_shared<ZeroInferRequest>(_initStructs, compiledModel, graph, config, _group_ordinal);
 }
 
 ov::SoPtr<ov::IRemoteTensor> ZeroDevice::createRemoteTensor(std::shared_ptr<ov::IRemoteContext> context,
