+++ src/plugins/intel_npu/src/backend/src/zero_infer_request.cpp
@@ -32,7 +32,7 @@ constexpr bool OUTPUT = false;
  * @param zeDescriptor The Level Zero specific structure used for comparison.
  */
 void check_level_zero_attributes_match(const IODescriptor& ioDescriptor,
-                                       const ZeroExecutor::ArgumentDescriptor& zeDescriptor) {
+                                       const IGraph::ArgumentDescriptor& zeDescriptor) {
     std::string zeDescriptorName = zeDescriptor.info.name;
 
     if (isStateInputName(zeDescriptorName)) {
@@ -159,14 +159,12 @@ std::optional<size_t> ZeroInferRequest::get_batch_size(const NetworkMetadata& me
 ZeroInferRequest::ZeroInferRequest(const std::shared_ptr<ZeroInitStructsHolder>& initStructs,
                                    const std::shared_ptr<const ICompiledModel>& compiledModel,
                                    const std::shared_ptr<IGraph>& graph,
-                                   const std::shared_ptr<const IExecutor>& executor,
                                    const Config& config,
                                    uint32_t group_ordinal)
     : SyncInferRequest(compiledModel, config),
       _initStructs(initStructs),
       _graph(graph),
-      _executorPtr(executor),
-      _executor(static_cast<const ZeroExecutor*>(_executorPtr.get())),
+      _executor(static_cast<const ZeroExecutor*>(_graph->get_executor().get())),
       _group_ordinal(group_ordinal),
       _config(config),
       _logger("ZeroInferRequest", config.get<LOG_LEVEL>()),
@@ -179,9 +177,8 @@ ZeroInferRequest::ZeroInferRequest(const std::shared_ptr<ZeroInitStructsHolder>&
                      _initStructs->getProfilingDdiTable()),
       _profilingQuery(0, _initStructs->getDevice(), _initStructs->getProfilingDdiTable()) {
     _logger.debug("ZeroInferRequest::ZeroInferRequest - SyncInferRequest");
-    const std::vector<ZeroExecutor::ArgumentDescriptor>& executorInputDescriptors = _executor->get_input_descriptors();
-    const std::vector<ZeroExecutor::ArgumentDescriptor>& executorOutputDescriptors =
-        _executor->get_output_descriptors();
+    const std::vector<IGraph::ArgumentDescriptor>& executorInputDescriptors = _graph->get_input_descriptors();
+    const std::vector<IGraph::ArgumentDescriptor>& executorOutputDescriptors = _graph->get_output_descriptors();
 
     auto proftype = config.get<PROFILING_TYPE>();
     if (proftype == ov::intel_npu::ProfilingType::INFER) {
@@ -288,7 +285,6 @@ void ZeroInferRequest::create_pipeline() {
     _pipeline = std::make_unique<Pipeline>(_config,
                                            _initStructs,
                                            _graph,
-                                           _executorPtr,
                                            _profilingPool,
                                            _profilingQuery,
                                            _npuProfiling,
@@ -345,8 +341,8 @@ void ZeroInferRequest::set_tensor_data(const std::shared_ptr<ov::ITensor> tensor
 
             OV_ITT_TASK_NEXT(ZERO_SET_TENSOR, "updateCommandList");
             _pipeline->updateCommandList(*tensorsData,
-                                         isInput ? _executor->get_input_descriptors().at(index).idx
-                                                 : _executor->get_output_descriptors().at(index).idx);
+                                         isInput ? _graph->get_input_descriptors().at(index).idx
+                                                 : _graph->get_output_descriptors().at(index).idx);
         }
     }
 }
@@ -377,9 +373,9 @@ void ZeroInferRequest::set_remote_tensor_data(const std::shared_ptr<ZeroRemoteTe
         _logger.debug("ZeroInferRequest::infer_async - update command list");
 
         OV_ITT_TASK_NEXT(ZERO_SET_REMOTE_TENSOR, "updateCommandList");
-        _pipeline->updateCommandList(*tensorsData,
-                                     isInput ? _executor->get_input_descriptors().at(index).idx
-                                             : _executor->get_output_descriptors().at(index).idx);
+        _pipeline->updateCommandList(
+            *tensorsData,
+            isInput ? _graph->get_input_descriptors().at(index).idx : _graph->get_output_descriptors().at(index).idx);
     }
 }
 
@@ -492,7 +488,7 @@ void ZeroInferRequest::set_tensors(const ov::Output<const ov::Node>& port,
                 if (_pipelineIsCreated) {
                     OV_ITT_TASK_NEXT(SET_TENSORS, "updateCommandList");
                     _pipeline->updateCommandList(*get_input_tensor_data(foundPort.idx, i),
-                                                 _executor->get_input_descriptors().at(foundPort.idx).idx,
+                                                 _graph->get_input_descriptors().at(foundPort.idx).idx,
                                                  i);
                 }
             }
@@ -747,15 +743,9 @@ std::vector<ov::ProfilingInfo> ZeroInferRequest::get_profiling_info() const {
     if (compilerType == ov::intel_npu::CompilerType::MLIR) {
         // For plugin compiler retreive raw profiling data from backend and delegate
         // processing to the compiler
-
-        OPENVINO_THROW("Not supported for now");
-
-        // const auto& graph = compiledModel.get_graph();
-        // const auto& compiler = compiledModel.get_compiler();
-        // const auto& blob = networkDesc->compiledNetwork;
-        // auto profData = get_raw_profiling_data();
-        // _logger.debug("InferRequest::get_profiling_info complete with compiler->process_profiling_output().");
-        // return compiler->process_profiling_output(profData, blob, compilerConfig);
+        auto profData = get_raw_profiling_data();
+        _logger.debug("InferRequest::get_profiling_info complete with compiler->process_profiling_output().");
+        return _graph->process_profiling_output(profData);
     } else {
         auto proftype = _config.get<PROFILING_TYPE>();
         if (proftype == ov::intel_npu::ProfilingType::INFER) {
