+++ src/plugins/intel_npu/src/compiler_adapter/include/cid_graph.hpp
@@ -9,33 +9,31 @@
 #include <ze_graph_ext.h>
 
 #include "igraph.hpp"
-#include "izero_compiler_in_driver.hpp"
-#include "npu.hpp"
-
+#include "izero_link.hpp"
 
 namespace intel_npu {
 
 class CidGraph final : public IGraph {
 public:
-    CidGraph(const std::shared_ptr<ILevelZeroCompilerInDriver>& apiAdapter,
+    CidGraph(const std::shared_ptr<IZeroLink>& zeroLink,
              ze_graph_handle_t graphHandle,
              NetworkMetadata metadata,
              const Config& config);
 
     CompiledNetwork export_blob() const override;
 
-    std::vector<ov::ProfilingInfo> process_profiling_output() const override;
+    std::vector<ov::ProfilingInfo> process_profiling_output(const std::vector<uint8_t>& profData) const override;
 
     void set_argument_value(uint32_t argi, const void* argv) const override;
 
-    void initialize() const override;
+    void initialize() override;
 
     ~CidGraph() override;
 
 private:
-    std::shared_ptr<ILevelZeroCompilerInDriver> _apiAdapter;
-    const Config _config;
+    std::shared_ptr<IZeroLink> _zeroLink;
 
+    const Config _config;
     Logger _logger;
 };
