+++ src/plugins/intel_npu/src/compiler_adapter/include/zero_link.hpp
@@ -10,13 +10,10 @@
 
 #include "intel_npu/utils/logger/logger.hpp"
 #include "intel_npu/utils/zero/zero_api.hpp"
-#include "izero_compiler_in_driver.hpp"
-#include "zero_executor.hpp"
-
+#include "izero_link.hpp"
+#include "zero_types.hpp"
 
 namespace intel_npu {
-
-using SerializedIR = std::pair<size_t, std::shared_ptr<uint8_t>>;
 
 #define NotSupportQuery(T) (T == ZE_GRAPH_EXT_VERSION_1_2)
 
@@ -46,30 +43,25 @@ using SerializedIR = std::pair<size_t, std::shared_ptr<uint8_t>>;
  * Adapter to use CiD through ZeroAPI
  */
 template <ze_graph_ext_version_t TableExtension>
-class LevelZeroCompilerInDriver final : public ILevelZeroCompilerInDriver {
+class ZeroLink final : public IZeroLink {
 public:
-    LevelZeroCompilerInDriver(ze_driver_handle_t driverHandle,
-                              ze_device_handle_t deviceHandle,
-                              ze_context_handle_t zeContext,
-                              ze_graph_dditable_ext_curr_t& graph_ddi_table_ext,
-                              ze_command_queue_npu_dditable_ext_curr_t& _commandQueueDdiTable,
-                              uint32_t group_ordinal);
-    LevelZeroCompilerInDriver(const LevelZeroCompilerInDriver&) = delete;
-    LevelZeroCompilerInDriver& operator=(const LevelZeroCompilerInDriver&) = delete;
-    ~LevelZeroCompilerInDriver();
+    ZeroLink(ze_driver_handle_t driverHandle,
+             ze_device_handle_t deviceHandle,
+             ze_context_handle_t zeContext,
+             ze_graph_dditable_ext_curr_t& graph_ddi_table_ext,
+             ze_command_queue_npu_dditable_ext_curr_t& _commandQueueDdiTable,
+             uint32_t group_ordinal);
+    ZeroLink(const ZeroLink&) = delete;
+    ZeroLink& operator=(const ZeroLink&) = delete;
+    ~ZeroLink();
 
-    uint32_t getSupportedOpsetVersion() const;
+    std::unordered_set<std::string> queryResultFromSupportedLayers(SerializedIR serializedIR,
+                                                                   const std::string& buildFlags) const override;
+    ze_graph_handle_t getGraphHandle(SerializedIR serializedIR,
+                                     const std::string& buildFlags,
+                                     const uint32_t& flags) const override;
 
-    ov::SupportedOpsMap query(const std::shared_ptr<const ov::Model>& model, const Config& config) const override;
-
-    ze_graph_handle_t compile(const std::shared_ptr<const ov::Model>& model, const Config& config) const override;
-
-    ze_result_t seriazlideIRModelAndCreateGraph(const std::shared_ptr<const ov::Model>& model,
-                                                const Config& config,
-                                                ze_device_graph_properties_t deviceGraphProperties,
-                                                ze_graph_handle_t& graphHandle) const;
-
-    ze_graph_handle_t parse(const std::vector<uint8_t>& network, const Config& config) const override;
+    ze_graph_handle_t getGraphHandle(const std::vector<uint8_t>& network) const override;
 
     template <ze_graph_ext_version_t T = TableExtension, std::enable_if_t<!NotSupportQuery(T), bool> = true>
     std::unordered_set<std::string> getQueryResultFromSupportedLayers(
@@ -77,22 +69,6 @@ public:
         ze_graph_query_network_handle_t& hGraphQueryNetwork) const;
 
     NetworkMetadata getNetworkMeta(ze_graph_handle_t graphHandle) const override;
-
-    /**
-     * @brief Serialize input / output information to string format.
-     * @details Format:
-     * --inputs_precisions="0:<input1Precision> [1:<input2Precision>]"
-     * --inputs_layouts="0:<input1Layout> [1:<input2Layout>]"
-     * --outputs_precisions="0:<output1Precision>"
-     * --outputs_layouts="0:<output1Layout>"
-     *
-     * For older compiler versions, the name of the inputs/outputs may be used instead of their indices.
-     *
-     * Since the layout information is no longer an important part of the metadata values when using the 2.0 OV
-     * API, the layout fields shall be filled with default values in order to assure the backward compatibility
-     * with the driver.
-     */
-    static std::string serializeIOInfo(const std::shared_ptr<const ov::Model>& model, const bool useIndices);
 
     _ze_result_t release(ze_graph_handle_t graphHandle) override;
 
@@ -102,11 +78,12 @@ public:
 
     void graphInitialie(ze_graph_handle_t graphHandle, const Config& config) const override;
 
-private:
-    SerializedIR serializeIR(const std::shared_ptr<const ov::Model>& model,
-                             ze_graph_compiler_version_info_t compilerVersion) const;
-    std::string serializeConfig(const Config& config, ze_graph_compiler_version_info_t& compilerVersion) const;
+    std::shared_ptr<IExecutor> createExecutor(ze_graph_handle_t graphHandle, const Config& config) const override;
 
+    std::tuple<std::vector<IGraph::ArgumentDescriptor>, std::vector<IGraph::ArgumentDescriptor>> getIODesc(
+        ze_graph_handle_t graphHandle) const override;
+
+private:
     template <ze_graph_ext_version_t T = TableExtension,
               typename std::enable_if_t<NotSupportArgumentMetadata(T), bool> = true>
     void getMetadata(ze_graph_dditable_ext_curr_t& graphDdiTableExt,
@@ -141,23 +118,20 @@ private:
 
     template <ze_graph_ext_version_t T = TableExtension,
               typename std::enable_if_t<SupportAPIGraphQueryNetworkV2(T), bool> = true>
-    ze_result_t seriazlideIRModelAndQueryNetworkCreateV2(const std::shared_ptr<const ov::Model>& model,
-                                                         const Config& config,
-                                                         ze_device_graph_properties_t deviceGraphProperties,
+    ze_result_t seriazlideIRModelAndQueryNetworkCreateV2(SerializedIR serializedIR,
+                                                         const std::string& buildFlags,
                                                          const ze_device_handle_t& _deviceHandle,
                                                          ze_graph_query_network_handle_t& hGraphQueryNetwork) const;
 
     // ext version >= 1.5, support API (pfnCreate2, pfnQueryNetworkCreate2, pfnQueryContextMemory)
     template <ze_graph_ext_version_t T = TableExtension,
               typename std::enable_if_t<SupportAPIGraphQueryNetworkV2(T), bool> = true>
-    std::unordered_set<std::string> queryImpl(const std::shared_ptr<const ov::Model>& model,
-                                              const Config& config) const;
+    std::unordered_set<std::string> queryImpl(SerializedIR serializedIR, const std::string& buildFlags) const;
 
     template <ze_graph_ext_version_t T = TableExtension,
               typename std::enable_if_t<SupportAPIGraphQueryNetworkV1(T), bool> = true>
-    ze_result_t seriazlideIRModelAndQueryNetworkCreateV1(const std::shared_ptr<const ov::Model>& model,
-                                                         const Config& config,
-                                                         ze_device_graph_properties_t deviceGraphProperties,
+    ze_result_t seriazlideIRModelAndQueryNetworkCreateV1(SerializedIR serializedIR,
+                                                         const std::string& buildFlags,
                                                          const ze_device_handle_t& _deviceHandle,
                                                          ze_graph_query_network_handle_t& hGraphQueryNetwork) const;
 
@@ -165,24 +139,20 @@ private:
     // pfnQueryNetworkGetSupportedLayers)
     template <ze_graph_ext_version_t T = TableExtension,
               typename std::enable_if_t<SupportAPIGraphQueryNetworkV1(T), bool> = true>
-    std::unordered_set<std::string> queryImpl(const std::shared_ptr<const ov::Model>& model,
-                                              const Config& config) const;
+    std::unordered_set<std::string> queryImpl(SerializedIR serializedIR, const std::string& buildFlags) const;
 
     // For ext version < 1.3
     template <ze_graph_ext_version_t T = TableExtension, typename std::enable_if_t<NotSupportQuery(T), bool> = true>
-    std::unordered_set<std::string> queryImpl(const std::shared_ptr<const ov::Model>& model,
-                                              const Config& config) const;
+    std::unordered_set<std::string> queryImpl(SerializedIR serializedIR, const std::string& buildFlags) const;
 
     template <ze_graph_ext_version_t T = TableExtension, typename std::enable_if_t<NotSupportGraph2(T), bool> = true>
-    ze_result_t createGraph(const ze_graph_format_t& format,
-                            const SerializedIR& serializedIR,
+    ze_result_t createGraph(SerializedIR serializedIR,
                             const std::string& buildFlags,
                             const uint32_t& flags,
                             ze_graph_handle_t* graph) const;
 
     template <ze_graph_ext_version_t T = TableExtension, typename std::enable_if_t<!NotSupportGraph2(T), bool> = true>
-    ze_result_t createGraph(const ze_graph_format_t& format,
-                            const SerializedIR& serializedIR,
+    ze_result_t createGraph(SerializedIR serializedIR,
                             const std::string& buildFlags,
                             const uint32_t& flags,
                             ze_graph_handle_t* graph) const;
