+++ src/plugins/intel_npu/src/compiler_adapter/src/cid_graph.cpp
@@ -4,38 +4,56 @@
 
 #include "cid_graph.hpp"
 
+#include "intel_npu/al/config/runtime.hpp"
+
 namespace intel_npu {
 
-CidGraph::CidGraph(const std::shared_ptr<ILevelZeroCompilerInDriver>& apiAdapter,
+CidGraph::CidGraph(const std::shared_ptr<IZeroLink>& zeroLink,
                    ze_graph_handle_t graphHandle,
                    NetworkMetadata metadata,
                    const Config& config)
-    : IGraph(static_cast<void*>(graphHandle), std::move(metadata)),
-      _apiAdapter(apiAdapter),
+    : IGraph(graphHandle, std::move(metadata)),
+      _zeroLink(zeroLink),
       _config(config),
       _logger("CidGraph", _config.get<LOG_LEVEL>()) {
-    initialize();
+    std::tie(_input_descriptors, _output_descriptors) = _zeroLink->getIODesc(_handle);
+
+    if (_config.get<CREATE_EXECUTOR>()) {
+        initialize();
+    } else {
+        _logger.info("Graph initialize is deferred from the \"Graph\" constructor");
+    }
 }
 
 CompiledNetwork CidGraph::export_blob() const {
-    return _apiAdapter->getCompiledNetwork(static_cast<ze_graph_handle_t>(_handle));
+    return _zeroLink->getCompiledNetwork(_handle);
 }
 
-std::vector<ov::ProfilingInfo> CidGraph::process_profiling_output() const {
-    OPENVINO_THROW("Profiling post-processing is not implemented.");
+std::vector<ov::ProfilingInfo> CidGraph::process_profiling_output(const std::vector<uint8_t>& profData) const {
+    OPENVINO_THROW("Profiling post-processing is not supported.");
 }
 
 void CidGraph::set_argument_value(uint32_t argi, const void* argv) const {
-    _apiAdapter->setArgumentValue(static_cast<ze_graph_handle_t>(_handle), argi, argv);
+    if (_zeroLink == nullptr) {
+        OPENVINO_THROW("Zero compiler adapter wasn't initialized");
+    }
+    _zeroLink->setArgumentValue(_handle, argi, argv);
 }
 
-void CidGraph::initialize() const {
-    _apiAdapter->graphInitialie(static_cast<ze_graph_handle_t>(_handle), _config);
+void CidGraph::initialize() {
+    if (_zeroLink) {
+        _logger.debug("Graph initialize start");
+
+        _zeroLink->graphInitialie(_handle, _config);
+        _executor = _zeroLink->createExecutor(_handle, _config);
+
+        _logger.debug("Graph initialize finish");
+    }
 }
 
 CidGraph::~CidGraph() {
     if (_handle != nullptr) {
-        auto result = _apiAdapter->release(static_cast<ze_graph_handle_t>(_handle));
+        auto result = _zeroLink->release(_handle);
 
         if (ZE_RESULT_SUCCESS == result) {
             _handle = nullptr;
