+++ src/plugins/intel_npu/src/plugin/src/compiled_model.cpp
@@ -23,13 +23,6 @@
 
 namespace {
 
-constexpr std::string_view NO_EXECUTOR_FOR_INFERENCE =
-    "Can't create infer request due to create executor failed! Only exports can be made.";
-
-constexpr std::string_view NO_EXECUTOR_FOR_INFERENCE_NODEVICE =
-    "Can't create infer request!\n"
-    "Please make sure that the device is available. Only exports can be made.";
-
 std::uint32_t hash(const intel_npu::CompiledNetwork& blob) {
     std::uint32_t result = 1171117u;
     for (const uint8_t* it = blob.data; it != blob.data + blob.size; ++it) {
@@ -61,9 +54,6 @@ CompiledModel::CompiledModel(const std::shared_ptr<const ov::Model>& model,
     initialize_properties();
     configure_stream_executors();
 
-    OV_ITT_TASK_NEXT(COMPILED_MODEL, "create_executor");
-    create_executor();
-
     OV_ITT_TASK_SKIP(COMPILED_MODEL);
 }
 
@@ -74,21 +64,12 @@ CompiledModel::~CompiledModel() {
 std::shared_ptr<ov::IAsyncInferRequest> CompiledModel::create_infer_request() const {
     OV_ITT_SCOPED_TASK(itt::domains::NPUPlugin, "CompiledModel::create_infer_request");
 
-    if (_executorPtr == nullptr && _device != nullptr) {
-        _executorPtr = _device->createExecutor(_graph, _config);
-    }
-
-    if (_executorPtr == nullptr) {
-        if (_device != nullptr) {
-            OPENVINO_THROW(NO_EXECUTOR_FOR_INFERENCE);
-        } else {
-            _logger.error("Can not find device!");
-            OPENVINO_THROW(NO_EXECUTOR_FOR_INFERENCE_NODEVICE);
-        }
+    if (!_config.get<CREATE_EXECUTOR>()) {
+        _graph->initialize();
     }
 
     const std::shared_ptr<SyncInferRequest>& syncInferRequest =
-        _device->createInferRequest(shared_from_this(), _graph, _executorPtr, _config);
+        _device->createInferRequest(shared_from_this(), _graph, _config);
     syncInferRequest->initialize_states();
 
     return std::make_shared<AsyncInferRequest>(syncInferRequest,
@@ -140,9 +121,9 @@ void CompiledModel::set_property(const ov::AnyMap& properties) {
     }
 
     _config.update(config);
-    if (_executorPtr != nullptr && config.find(ov::workload_type.name()) != config.end()) {
+    if (config.find(ov::workload_type.name()) != config.end()) {
         const auto workloadType = properties.at(ov::workload_type.name()).as<ov::WorkloadType>();
-        _executorPtr->setWorkloadType(workloadType);
+        _graph->setWorkloadType(workloadType);
     }
 }
 
@@ -153,10 +134,6 @@ ov::Any CompiledModel::get_property(const std::string& name) const {
     }
 
     OPENVINO_THROW("Unsupported property ", name);
-}
-
-const std::shared_ptr<IGraph>& CompiledModel::get_graph() const {
-    return _graph;
 }
 
 const Config& CompiledModel::get_config() const {
@@ -183,6 +160,10 @@ void CompiledModel::configure_stream_executors() {
     set_task_executor(std::move(task_executor));
     const auto executorId = _graph->get_metadata().name + "_NPUResultExecutor";
     _resultExecutor = ov::threading::executor_manager()->get_executor(executorId);
+}
+
+const NetworkMetadata& CompiledModel::get_network_metadata() const {
+    return _graph->get_metadata();
 }
 
 void CompiledModel::initialize_properties() {
@@ -361,20 +342,6 @@ void CompiledModel::initialize_properties() {
         if (std::get<0>(property.second)) {
             _supportedProperties.emplace_back(property.first, std::get<1>(property.second));
         }
-    }
-}
-
-void CompiledModel::create_executor() {
-    if (_config.get<CREATE_EXECUTOR>()) {
-        _logger.info("Creating the executor inside the \"CompiledModel\" constructor");
-
-        // If no device has been defined, the executor shall keep the default value of "nullptr". In this scenario,
-        // only export operations will be allowed
-        if (_device != nullptr) {
-            _executorPtr = _device->createExecutor(_graph, _config);
-        }
-    } else {
-        _logger.info("Executor will not be created inside the \"CompiledModel\" constructor");
     }
 }
